third_party/WebKit/LayoutTests/platform/win/http/tests/w3c/webperf/approved/navigation-timing/html/test_navigation_attributes_exist-expected.txt

This is a testharness.js-based test.
PASS window.performance is defined 
PASS window.performance.navigation is defined 
PASS window.performance.navigation.type is defined. 
PASS window.performance.navigation.redirectCount is defined. 
Harness: the test ran to completion.
